markerClick opens modal
