+++ .idea/workspace.xml
@@ -5,11 +5,7 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="0f63a5fc-013b-4ba2-9e36-465b6c3bdbcd" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/App.js" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/App.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/package-lock.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/package.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/screens/Login.js" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/screens/Login.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/frontend/components/Info_card.js" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/components/Info_card.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/frontend/screens/Map.js" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/screens/Map.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
